cookiecutter: Add django.db.backends logger
"For performance reasons, SQL logging is only enabled when settings.DEBUG is set to True,
regardless of the logging level or handlers that are installed"

Set level to "INFO" if in development, to not spam developers. Developers can opt in with LOG_LEVEL=DEBUG.

Set level to "DEBUG" in production. Administrators can opt in with DEBUG=True.

Although we test for other boolean environment variables with "VAR" in os.environ,
we especially don't want to accidentally enable DEBUG with DEBUG=False. So, we use a different test.